edited rake tasks to deal with strange bugs

+++ lib/shared_database/tasks/db/drop.rake
@@ -5,12 +5,23 @@ Rake::Task["db:drop"].clear if Rake::Task.task_defined?("db:drop")
 namespace :db do
     desc "(Shared_Database) Drop the database"
     task :drop do
-        SharedDatabase.connect_db
-        if SharedDatabase.db_config["adapter"] != "sqlite3"
-            ActiveRecord::Base.connection.drop_database(db_config["database"])
-        else  # sqlite3 drop db is basically deleting the file
+        case SharedDatabase.db_config['adapter']
+        when /mysql/
+            SharedDatabase.connect_public_db
+            ActiveRecord::Base.connection.drop_database(SharedDatabase.db_config["database"])
+            puts "MySQL database deleted."
+            abort  # return does not exist
+        when /^sqlite/
             File.delete(SharedDatabase.db_config["database"])
+            puts "Sqlite database deleted."
+            abort
+        when 'postgresql'
+            SharedDatabase.connect_public_db
+            ActiveRecord::Base.connection.drop_database(SharedDatabase.db_config["database"])
+            puts "Postgresql database deleted."
+            abort
         end
-        puts "Database deleted."
+
+        puts "Database failed to delete..."
     end
 end
